Add parameterize feature of placeholders
Parse and recognize the value's placeholders in condition-expression
