+++ lib/aws-dynamodb.js
@@ -1,6 +1,9 @@
 (function() {
     var fs = require('fs');
     var awscli = require("../lib/awscli");
+    var DynamoDbScanStatement = require('../lib/dynamodb-scan-statement.js');
+    var DynamoDbQueryStatement = require('../lib/dynamodb-query-statement.js');
+
     //awscli.setDebug();
     var DynamoDB = null;
     awscli.dynamodb = awscli.dynamodb || {};
@@ -68,29 +71,18 @@
         callback(null, create);
     };
 
-    /**
-     * Scan dynamodb table
-     * @param {DynamodbStatement} statement DynamodbStatement
-     * @param {Function} callback callback to receive result set.
-     * @returns {undefined}
-     */
-    awscli.dynamodb.runScanStatemnt = function(statement, callback) {
-        var param = statement.getScanParameter();
-        DynamoDB.scan(param, callback);
+    awscli.dynamodb.ScanStatement = function(sql) {
+        var statement = new DynamoDbScanStatement(sql);
+        statement.dynamodb = DynamoDB;
+        return statement;
+    };
+    awscli.dynamodb.QueryStatement = function(sql) {
+        var statement = new DynamoDbQueryStatement(sql);
+        statement.dynamodb = DynamoDB;
+        return statement;
     };
 
-    /**
-     * Query dynamodb table
-     * @param {DynamodbStatement} statement DynamodbStatement
-     * @param {Function} callback callback to receive result set.
-     * @returns {undefined}
-     */
-    awscli.dynamodb.runQueryStatemnt = function(statement, callback) {
-        var param = statement.getQueryParameter();
-        DynamoDB.query(param, callback);
-    };
-
-    var parsers = require("../lib/aws-dynamodb-expr-parsers");
+    var parsers = require("../lib/dynamodb-sqlish-parser");
     awscli.dynamodb.isKeyword = parsers.isKeyword;
 
     module.exports = awscli.dynamodb;
