+++ lib/bnf.js
@@ -8,14 +8,120 @@ BNF.literal = function(value) { return { role: "lit", value: value }; };
 BNF.lex = function(value) { return { role: "lex", value: value }; };
 BNF.ident = BNF.lex("IDENT");
 BNF.numlit = BNF.lex("NUMLIT");
-BNF.strlit = BNF.lex("STRLIT");
+BNF.strlitDQ = BNF.lex("STRLIT-DQ");
+BNF.strlitSQ = BNF.lex("STRLIT-SQ");
 BNF.comma = BNF.literal(",");
 
-function log() {
-    var arg = Array.from(arguments);
-    var indent = arg.splice(0,1);
-    console.log("  ".repeat(indent) + arg.join(" "));
+var log = function() {};
+BNF.setDebug = function(debug) {
+    if(debug) {
+        log = function() {
+            var arg = Array.from(arguments);
+            var indent = arg.splice(0,1);
+            console.log("  ".repeat(indent) + arg.join(" "));
+        }
+    } else {
+        log = function() {};
+    }
+};
+
+BNF.ParseResult = function() { };
+BNF.ParseResult.prototype.existsTerm = function(name) {
+    var term = this.getTerm(name);
+    if(term !== false) {
+        return term.match;
+    }
+    return false;
+};
+BNF.ParseResult.prototype.getTerm = function(name) {
+    var n = this.terms.length;
+    for(var i = 0; i < n; i++) {
+        if(this.terms[i].term === name) {
+            log(0, "[BNF.ParseResult.getTerm]", name, "found");
+            return this.terms[i];
+        }
+    }
+    log(0, "[BNF.ParseResult.getTerm]", name, "NOT found");
+    return false;
+};
+function strUnescape(s) {
+    var ss = "";
+    var len = s.length;
+    for(var i = 0; i < len; i++) {
+        var c = s.charAt(i);
+        var cc = c.charCodeAt(0);
+        switch(c) {
+            case "\t": ss += "\\t"; break;
+            case "\r": ss += "\\r"; break;
+            case "\n": ss += "\\n"; break;
+            default:
+                if(0x20 <= cc && cc <= 0x7f || cc > 0xff) {
+                    ss += c;
+                } else if(cc <= 0xff) {
+                    ss += "\\x" + parseInt(cc, 16);
+                }
+                break;
+        }
+    }
+    return ss;
 }
+function unescapeDQ(s) {
+    s = strUnescape(s);
+    var ss = "";
+    var len = s.length;
+    for(var i = 0; i < len; i++) {
+        var c = s.charAt(i);
+        var cc = c.charCodeAt(0);
+        if(c == "\"") {
+            ss += "\\\"";
+        } else {
+            ss += c;
+        }
+    }
+    return ss;
+}
+function unescapeSQ(s) {
+    s = strUnescape(s);
+    var ss = "";
+    var len = s.length;
+    for(var i = 0; i < len; i++) {
+        var c = s.charAt(i);
+        var cc = c.charCodeAt(0);
+        if(c == "'") {
+            ss += "\\'";
+        } else {
+            ss += c;
+        }
+    }
+    return ss;
+}
+BNF.ParseResult.prototype.getTermsList = function(indent) {
+    indent = indent || 0;
+    var s = []
+    this.terms.forEach(function(term) {
+        log(indent, term.term, term.terms.length);
+        if(term.terms.length == 0) {
+            switch(term.lex.getType()) {
+            case "STRLIT-DQ":
+                s.push("\"" + unescapeDQ(term.lex.getTerm()) + "\"");
+                break;
+            case "STRLIT-SQ":
+                s.push("'" + unescapeSQ(term.lex.getTerm()) + "'");
+                break;
+            default:
+                s.push(term.lex.getTerm());
+                break;
+            }
+        } else {
+            term.getTermsList(indent+1).forEach(function(term) {
+                s.push(term);
+            });
+        }
+    });
+    log(indent, s.join('|'));
+    return s;
+};
+
 BNF.prototype.parse = function(source) {
     log(0, "source:", source);
     var rawLexList = lexana.parseLexicalElements(source);
@@ -31,28 +137,37 @@ BNF.prototype.parse = function(source) {
             {
                 lex.setTerm(lex.getTerm() + lex2.getTerm());
                 rawLexList[i + 1] = null;
+            } else if(lex2 != null && lex.getType() === "PUNCT" &&
+                     lex.getTerm() === ":" && lex2.getType() === "IDENT")
+            {
+                lex.setType("IDENT");
+                lex.setTerm(lex.getTerm() + lex2.getTerm());
+                rawLexList[i + 1] = null;
             }
             lexList.push(lex);
-            log(0, JSON.stringify(lex));
         }
     }
-    return this.parseSentence(this._root, lexList, 0, 0);
+    var result = this.parseSentence(this._root, lexList, 0, 0);
+    log(0, JSON.stringify(result, null, "  "));
+    BNF.logResult(result);
+    return result;
 };
 
 BNF.prototype.parseSentence = function(root, lexList, lexIndex, indent) {
     log(indent, "BNF:", root, "(at", lexIndex, ")");
     var parser = {};
     if(!(root in this._bnf)) {
-        throw new Err(root + " is not declared in BNF");
+        throw new Error(root + " is not declared in BNF");
     }
     var declaration = this._bnf[root];
     var nDecl = declaration.length;
-    var result = {
-        match: false,
-        name: root,
-        lexCount: 0,
-        terms: []
-    };
+
+    var result = new BNF.ParseResult();
+    result.match = false;
+    result.term = root;
+    result.lexCount = 0;
+    result.terms = [];
+
     for(var iDecl = 0; iDecl < nDecl; iDecl++) {
         var termList = declaration[iDecl];
         var termResult = this.parseTermList(termList, lexList, lexIndex, indent + 1);
@@ -68,53 +183,54 @@ BNF.prototype.parseSentence = function(root, lexList, lexIndex, indent) {
 
 BNF.prototype.parseTermList = function(termList, lexList, lexIndex, indent) {
     var nTerm = termList.length;
-    var matchCount = 0;
-    var result = {
-        match: true,
-        lexCount: 0,
-        terms: []
-    };
 
-    for(var iTerm = 0; iTerm < nTerm && lexIndex < lexList.length; iTerm++) {
-        var term = termList[iTerm];
-        var termType = typeof(term);
-        var termResult = {
-            term: term,
-            match: null,
-            lex: null,
-            lexCount: 0,
-            optional: false,
-            terms: []
-        };
-        if(termType != "string" && termType !== "object" || Array.isArray(term)) {
-            throw new Error("Illegal BNF definition at " + JSON.stringify(term));
+    var result = new BNF.ParseResult();
+    result.match = true;
+    result.lexCount = 0;
+    result.terms = [];
+
+    for(var iTerm = 0; iTerm < nTerm; iTerm++) {
+        var termResult = new BNF.ParseResult();
+        termResult.term = termList[iTerm];
+        termResult.match = false;
+        termResult.lex = null;
+        termResult.optional = false;
+        termResult.terms = [];
+
+        var termType = typeof(termResult.term);
+        if(termType != "string" && termType !== "object" || Array.isArray(termResult.term)) {
+            throw new Error("Illegal BNF definition at " + JSON.stringify(termResult.term));
         }
-        if(typeof(term) === "string") {
-            termResult.optional = (term.match(/^\[.*\]$/) ? true : false);
-            term = term.replace(/^\[(.*)\]$/, "$1");
-            var subResult = this.parseSentence(term, lexList, lexIndex, indent);
-            termResult.match = subResult.match;
-            termResult.terms = subResult.terms;
-            if(termResult.match) {
-                termResult.lexCount = subResult.lexCount;
+        if(termType === "string") {
+            termResult.optional = (termResult.term.match(/^\[.*\]$/) ? true : false);
+            termResult.term = termResult.term.replace(/^\[(.*)\]$/, "$1");
+            if(lexIndex < lexList.length) {
+                var subResult = this.parseSentence(termResult.term, lexList, lexIndex, indent);
+                termResult.match = subResult.match;
+                termResult.terms = subResult.terms;
+                if(termResult.match) {
+                    result.lexCount += subResult.lexCount;
+                    lexIndex += subResult.lexCount;
+                }
             }
         } else {
-            var lex = lexList[lexIndex];
-            var lexValue = null;
-            switch(term.role) {
-                case "lit": lexValue = lex.getTerm(); break;
-                case "lex": lexValue = lex.getType(); break;
-            }
-            termResult.lex = lex;
-            termResult.match = (lexValue === term.value);
-            if(termResult.match) {
-                termResult.lexCount = 1;
-                log(indent, "match: ", JSON.stringify(termResult.lex),
-                        "( at " + lexIndex + ")");
+            if(lexIndex < lexList.length) {
+                termResult.lex = lexList[lexIndex];
+                var lexValue = null;
+                switch(termResult.term.role) {
+                    case "lit": lexValue = termResult.lex.getTerm(); break;
+                    case "lex": lexValue = termResult.lex.getType(); break;
+                }
+
+                termResult.match = (lexValue.toUpperCase() === termResult.term.value.toUpperCase());
+                if(termResult.match) {
+                    result.lexCount++;
+                    lexIndex++;
+                    log(indent, "match: ", JSON.stringify(termResult.lex),
+                            "( at " + lexIndex + ")");
+                }
             }
         }
-        result.lexCount += termResult.lexCount;
-        lexIndex += termResult.lexCount;
         result.terms.push(termResult);
         if(!termResult.optional && !termResult.match) {
             result.match = false;
@@ -122,11 +238,12 @@ BNF.prototype.parseTermList = function(termList, lexList, lexIndex, indent) {
         }
     }
     return result;
-}
+};
 
 BNF.logResult = function(parseResult) {
     BNF._logResult(parseResult.terms);
-}
+};
+
 BNF._logResult = function(parseResult, indent) {
     indent = indent || 0;
     parseResult.forEach(function(e) {
@@ -138,5 +255,5 @@ BNF._logResult = function(parseResult, indent) {
                     JSON.stringify(e.lex.getTerm()));
         }
     });
-}
+};
 module.exports = BNF;
