+++ lib/lexana.js
@@ -84,14 +84,15 @@ module.exports.parseLexicalElements = function(source) {
         token.fixTerm();
         tokenList.push(token);
         toMode("");
+        token = null;
     };
     var parserError = function(message) {
-        return {
-            "lineNumber": lineNum,
-            "column": columnPos,
-            "message": message,
-            "tokenList" : tokenList
-        };
+        var err = new Error(message);
+        err.lineNumber = lineNum;
+        err.column = columnPos;
+        err.message = message;
+        err.tokenList = tokenList;
+        return err;
     };
     var ungetChar = function() {
         --i;
@@ -99,7 +100,7 @@ module.exports.parseLexicalElements = function(source) {
     };
     var refNextChar = function() {
         if(i + 1 >= source.length) {
-            throw parserError("unexpected terminate.");
+            throw parserError("unexpected terminate");
         }
         return source.charAt(i + 1);
     };
@@ -124,7 +125,12 @@ module.exports.parseLexicalElements = function(source) {
             ++i;
             ++columnPos;
         }
-        endToken();
+        if(token != null) {
+            if( mode === "WS-BLKCMT" || mode === "STRLIT") {
+                throw parserError("unexpected terminate");
+            }
+            endToken();
+        }
     }({
         "" : function() {
             token = new Token();
@@ -244,31 +250,4 @@ module.exports.parseLexicalElements = function(source) {
         }
     }));
     return tokenList;
-};
-//
-// Ecma BNF classes
-// http://tomcopeland.blogs.com/EcmaScript.html
-//
-var EcmaScriptBnfElement = function() {};
-var Literal = function() {};
-Literal.prototype = new EcmaScriptBnfElement();
-var DecimalLiteral = function() { };
-DecimalLiteral.prototype = new EcmaScriptBnfElement();
-module.exports.parseWord = function(lextoks) {
-    var tokenList = [];
-    var i = 0;
-    var DFL = 0;
-    var STRS = 1;
-    var STRD = 2;
-    var NUM = 3;
-    var mode = DFL;
-    while(i < lextoks.length) {
-        var lextok = lextoks[i++];
-        if(mode == DFL) {
-        } else if(mode == STRS) {
-        } else if(mode == STRD) {
-        } else if(mode == NUM) {
-        } else {
-        }
-    }
 };
