+++ package.json
@@ -5,6 +5,7 @@
   "directries": {
     "lib": "./lib"
   },
+  "main": "./index.js",
   "bin": {
     "aws-lambda-get": "./bin/aws-lambda-get.sh",
     "aws-lambda-create": "./bin/aws-lambda-create.sh",
@@ -40,6 +41,7 @@
   "dependencies": {
     "aws-sdk": "^2.3.17",
     "camelcase": "^3.0.0",
+    "clone": "^2.1.1",
     "hash-arg": "0.1.0",
     "home-dir": "^1.0.0",
     "ini": "^1.3.4",
